pyproject.toml
[project]
name = "homework"
version = "0.1.0"
description = "School location server"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "beanie<2",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "mongomock-motor",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
